backend/requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.18.0
pydantic>=1.8
ultralytics>=8.0.0
torch>=1.12.0
torchvision>=0.13.0
Pillow>=9.0.0
numpy>=1.21.0
